iox-#1345 Move code from header into cpp and inl files
