+++ iceoryx_hoofs/include/iceoryx_hoofs/log_ng/platform/logger.hpp
@@ -1,5 +1,4 @@
-// Copyright (c) 2019 by Robert Bosch GmbH. All rights reserved.
-// Copyright (c) 2021 - 2022 by Apex.AI Inc. All rights reserved.
+// Copyright (c) 2022 by Apex.AI Inc. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -18,6 +17,7 @@
 #ifndef IOX_HOOFS_PLATFORM_LOG_LOGGER_HPP
 #define IOX_HOOFS_PLATFORM_LOG_LOGGER_HPP
 
+#include "iceoryx_hoofs/log_ng/platform_building_blocks/console_logger.hpp"
 #include "iceoryx_hoofs/log_ng/platform_building_blocks/logger.hpp"
 
 namespace iox
@@ -25,7 +25,6 @@ namespace iox
 namespace platform
 {
 using LogLevel = pbb::LogLevel;
-using pbb::asStringLiteral;
 using pbb::logLevelFromEnvOr;
 
 using Logger = pbb::Logger<pbb::ConsoleLogger>;
