+++ iceoryx_hoofs/include/iceoryx_hoofs/log_ng/platform_building_blocks/logcommon.hpp
@@ -1,0 +1,68 @@
+// Copyright (c) 2019 by Robert Bosch GmbH. All rights reserved.
+// Copyright (c) 2021 - 2022 by Apex.AI Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#ifndef IOX_HOOFS_PLATFORM_BUILDING_BLOCKS_LOGCOMMON_HPP
+#define IOX_HOOFS_PLATFORM_BUILDING_BLOCKS_LOGCOMMON_HPP
+
+#include <cstdint>
+
+namespace iox
+{
+namespace pbb
+{
+enum class LogLevel : uint8_t
+{
+    OFF = 0,
+    FATAL,
+    ERROR,
+    WARN,
+    INFO,
+    DEBUG,
+    TRACE // this could be used instead of commenting the code; with MINIMAL_LOG_LEVEL set to DEBUG, the compiler would
+    // optimize this out and there wouldn't be a performance hit
+};
+
+constexpr const char* LOG_LEVEL_COLOR[] = {
+    "",                 // nothing
+    "\033[0;1;97;41m",  // bold bright white on red
+    "\033[0;1;31;103m", // bold red on light yellow
+    "\033[0;1;93m",     // bold bright yellow
+    "\033[0;1;92m",     // bold bright green
+    "\033[0;1;96m",     // bold bright cyan
+    "\033[0;1;36m",     // bold cyan
+};
+
+constexpr const char* LOG_LEVEL_TEXT[] = {
+    "[ Off ]", // nothing
+    "[Fatal]", // bold bright white on red
+    "[Error]", // bold red on light yellow
+    "[Warn ]", // bold bright yellow
+    "[Info ]", // bold bright green
+    "[Debug]", // bold bright cyan
+    "[Trace]", // bold cyan
+};
+
+struct LogBuffer
+{
+    const char* buffer;
+    uint64_t writeIndex;
+};
+
+} // namespace pbb
+} // namespace iox
+
+#endif // IOX_HOOFS_PLATFORM_BUILDING_BLOCKS_CONSOLE_LOGGER_HPP
